requirements.txt
flask
gunicorn
torch
transformers
python-docx
PyPDF2
